Implement basic `run` feature
Using `2a-emulator run ARGS` the machine can be emulated for a number of clock cycles. The results of the run will be printed. Inputs can be given as arguments on the command line. See `2a-emulator run --help` for more information. This feature will at some point replace the testing capabilities the program has at the moment.