Reduce expectations for .ORG-instruction
This drops the failing test in favour of one that's actually passing.
This means `.ORG n` will not be able to decrement the byte counter. This
is not really bad, I find, since this prevents messing with your own
program by wrongly setting `.ORG n` somewhere.

But this can lead to diverging behaviour of the emulator from the real
machine. Because of this, the emulator will throw a hard error.

See #3.